+++ core/src/math/BitPacking.inl
@@ -57,10 +57,10 @@ uint64_t BitPacking::byteSwap64( uint64_t x )
 
 void BitPacking::byteSwap16( Array1DView< uint16_t > view16 )
 {
-    if( view16.packed( ) )
+    if( view16.packed() )
     {
-        int nWords64 = view16.size( ) / 64;
-        Array1DView< uint64_t > view64( view16.pointer( ), nWords64 );
+        int nWords64 = static_cast< int >( view16.size() ) / 64;
+        Array1DView< uint64_t > view64( view16.pointer(), nWords64 );
         for( int i = 0; i < nWords64; ++i )
         {
             uint64_t a = view64[ i ];
@@ -70,19 +70,19 @@ void BitPacking::byteSwap16( Array1DView< uint16_t > view16 )
         }
 
         // Input length may not be a multiple of 4.
-        if( nWords64 * 4 < view16.size( ) )
+        if( nWords64 * 4 < view16.size() )
         {
             // Could in theory do a byteSwap32() and then a byteSwap16()
             // if there were 3 elements left, but it's totally not worth it.
-            for( int i = nWords64 * 4; i < view16.size( ); ++i )
+            for( int i = nWords64 * 4; i < view16.size(); ++i )
             {
-                byteSwap16( view16[ view16.size( ) - 1 ] );
+                byteSwap16( view16[ view16.size() - 1 ] );
             }
         }
     }
     else
     {
-        for( int i = 0; i < view16.size( ); ++i )
+        for( int i = 0; i < view16.size(); ++i )
         {
             view16[ i ] = byteSwap16( view16[ i ] );
         }
